apps/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
bcrypt==4.0.1
boto3
redis
httpx